+++ app/src/main/java/com/termux/app/terminal/TermuxSessionsListViewController.java
@@ -20,7 +20,7 @@ import androidx.core.content.ContextCompat;
 
 import com.termux.R;
 import com.termux.app.TermuxActivity;
-import com.termux.shared.termux.shell.TermuxSession;
+import com.termux.shared.termux.shell.command.runner.terminal.TermuxSession;
 import com.termux.shared.theme.ThemeUtils;
 import com.termux.terminal.TerminalSession;
