+++ app/src/main/java/com/termux/app/utils/PluginUtils.java
@@ -10,15 +10,15 @@ import androidx.annotation.Nullable;
 import com.termux.R;
 import com.termux.shared.activities.ReportActivity;
 import com.termux.shared.file.FileUtils;
-import com.termux.shared.file.TermuxFileUtils;
-import com.termux.shared.models.ResultConfig;
-import com.termux.shared.models.ResultData;
-import com.termux.shared.models.errors.Errno;
-import com.termux.shared.models.errors.Error;
+import com.termux.shared.termux.file.TermuxFileUtils;
+import com.termux.shared.shell.command.result.ResultConfig;
+import com.termux.shared.shell.command.result.ResultData;
+import com.termux.shared.errors.Errno;
+import com.termux.shared.errors.Error;
 import com.termux.shared.notification.NotificationUtils;
 import com.termux.shared.termux.notification.TermuxNotificationUtils;
 import com.termux.shared.termux.settings.preferences.TermuxPreferenceConstants;
-import com.termux.shared.shell.ResultSender;
+import com.termux.shared.shell.command.result.ResultSender;
 import com.termux.shared.shell.ShellUtils;
 import com.termux.shared.android.AndroidUtils;
 import com.termux.shared.termux.TermuxConstants;
@@ -29,7 +29,7 @@ import com.termux.shared.termux.settings.preferences.TermuxPreferenceConstants.T
 import com.termux.shared.settings.properties.SharedProperties;
 import com.termux.shared.models.ReportInfo;
 import com.termux.shared.termux.settings.properties.TermuxPropertyConstants;
-import com.termux.shared.models.ExecutionCommand;
+import com.termux.shared.shell.command.ExecutionCommand;
 import com.termux.app.models.UserAction;
 import com.termux.shared.data.DataUtils;
 import com.termux.shared.markdown.MarkdownUtils;
