+++ app/src/main/java/com/termux/filepicker/TermuxFileReceiverActivity.java
@@ -12,9 +12,9 @@ import androidx.annotation.NonNull;
 import com.termux.R;
 import com.termux.shared.data.DataUtils;
 import com.termux.shared.data.IntentUtils;
-import com.termux.shared.net.UriUtils;
+import com.termux.shared.net.uri.UriUtils;
 import com.termux.shared.interact.MessageDialogUtils;
-import com.termux.shared.models.net.UriScheme;
+import com.termux.shared.net.uri.UriScheme;
 import com.termux.shared.termux.interact.TextInputDialogUtils;
 import com.termux.shared.termux.TermuxConstants;
 import com.termux.shared.termux.TermuxConstants.TERMUX_APP.TERMUX_SERVICE;
