+++ termux-shared/src/main/java/com/termux/shared/crash/CrashHandler.java
@@ -7,7 +7,7 @@ import androidx.annotation.NonNull;
 import com.termux.shared.file.FileUtils;
 import com.termux.shared.logger.Logger;
 import com.termux.shared.markdown.MarkdownUtils;
-import com.termux.shared.models.errors.Error;
+import com.termux.shared.errors.Error;
 import com.termux.shared.android.AndroidUtils;
 
 import java.nio.charset.Charset;
