+++ termux-shared/src/main/java/com/termux/shared/errors/Error.java
@@ -1,4 +1,4 @@
-package com.termux.shared.models.errors;
+package com.termux.shared.errors;
 
 import android.content.Context;
