+++ termux-shared/src/main/java/com/termux/shared/shell/command/result/ResultData.java
@@ -1,12 +1,12 @@
-package com.termux.shared.models;
+package com.termux.shared.shell.command.result;
 
 import androidx.annotation.NonNull;
 
 import com.termux.shared.data.DataUtils;
 import com.termux.shared.logger.Logger;
 import com.termux.shared.markdown.MarkdownUtils;
-import com.termux.shared.models.errors.Errno;
-import com.termux.shared.models.errors.Error;
+import com.termux.shared.errors.Errno;
+import com.termux.shared.errors.Error;
 
 import java.io.Serializable;
 import java.util.ArrayList;
