+++ termux-shared/src/main/java/com/termux/shared/shell/command/result/ResultSenderErrno.java
@@ -1,4 +1,6 @@
-package com.termux.shared.models.errors;
+package com.termux.shared.shell.command.result;
+
+import com.termux.shared.errors.Errno;
 
 /** The {@link Class} that defines ResultSender error messages and codes. */
 public class ResultSenderErrno extends Errno {
