+++ termux-shared/src/main/java/com/termux/shared/theme/ThemeUtils.java
@@ -3,8 +3,6 @@ package com.termux.shared.theme;
 import android.content.Context;
 import android.content.res.Configuration;
 
-import com.termux.shared.models.theme.NightMode;
-
 public class ThemeUtils {
 
     /**
